Fix: Move classes imported from ergebnis/factory-girl-definition
