+++ test/Fixture/Definition/ExtendsAbstractDefinition/UserDefinition.php
@@ -3,18 +3,18 @@
 declare(strict_types=1);
 
 /**
- * Copyright (c) 2017-2020 Andreas Möller
+ * Copyright (c) 2020 Andreas Möller
  *
  * For the full copyright and license information, please view
  * the LICENSE.md file that was distributed with this source code.
  *
- * @see https://github.com/ergebnis/factory-girl-definition
+ * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryGirl\Definition\Test\Fixture\Definition\ExtendsAbstractDefinition;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definition\ExtendsAbstractDefinition;
 
-use Ergebnis\FactoryGirl\Definition\AbstractDefinition;
-use Ergebnis\FactoryGirl\Definition\Test\Fixture\Entity;
+use Ergebnis\FactoryBot\Definition\AbstractDefinition;
+use Ergebnis\FactoryBot\Test\Fixture\Entity;
 use FactoryGirl\Provider\Doctrine\FixtureFactory;
 
 final class UserDefinition extends AbstractDefinition
