+++ test/Unit/Definition/AbstractDefinitionTest.php
@@ -3,19 +3,18 @@
 declare(strict_types=1);
 
 /**
- * Copyright (c) 2017-2020 Andreas Möller
+ * Copyright (c) 2020 Andreas Möller
  *
  * For the full copyright and license information, please view
  * the LICENSE.md file that was distributed with this source code.
  *
- * @see https://github.com/ergebnis/factory-girl-definition
+ * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryGirl\Definition\Test\Unit;
+namespace Ergebnis\FactoryBot\Test\Unit\Definition;
 
-use Ergebnis\FactoryGirl\Definition\AbstractDefinition;
-use Ergebnis\FactoryGirl\Definition\FakerAwareDefinition;
-use Ergebnis\FactoryGirl\Definition\Test\Fixture;
+use Ergebnis\FactoryBot\Definition\AbstractDefinition;
+use Ergebnis\FactoryBot\Definition\FakerAwareDefinition;
 use Ergebnis\Test\Util\Helper;
 use Faker\Generator;
 use PHPUnit\Framework;
@@ -23,7 +22,7 @@ use PHPUnit\Framework;
 /**
  * @internal
  *
- * @covers \Ergebnis\FactoryGirl\Definition\AbstractDefinition
+ * @covers \Ergebnis\FactoryBot\Definition\AbstractDefinition
  */
 final class AbstractDefinitionTest extends Framework\TestCase
 {
@@ -36,7 +35,7 @@ final class AbstractDefinitionTest extends Framework\TestCase
 
     public function testFakerThrowsBadMethodCallExceptionIfDefinitionHasNotBeenProvidedWithFaker(): void
     {
-        $definition = new Fixture\Definition\ExtendsAbstractDefinition\UserDefinition();
+        $definition = new \Ergebnis\FactoryBot\Test\Fixture\Definition\ExtendsAbstractDefinition\UserDefinition();
 
         $this->expectException(\BadMethodCallException::class);
         $this->expectExceptionMessage(\sprintf(
@@ -51,7 +50,7 @@ final class AbstractDefinitionTest extends Framework\TestCase
     {
         $faker = new Generator();
 
-        $definition = new Fixture\Definition\ExtendsAbstractDefinition\UserDefinition();
+        $definition = new \Ergebnis\FactoryBot\Test\Fixture\Definition\ExtendsAbstractDefinition\UserDefinition();
 
         $definition->provideWith($faker);
