--- test/Unit/ExampleTest.php
@@ -1,37 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * Copyright (c) 2020 Andreas Möller
- *
- * For the full copyright and license information, please view
- * the LICENSE.md file that was distributed with this source code.
- *
- * @see https://github.com/ergebnis/factory-bot
- */
-
-namespace Ergebnis\FactoryBot\Test\Unit;
-
-use Ergebnis\FactoryBot\Example;
-use Ergebnis\Test\Util\Helper;
-use PHPUnit\Framework;
-
-/**
- * @internal
- *
- * @covers \Ergebnis\FactoryBot\Example
- */
-final class ExampleTest extends Framework\TestCase
-{
-    use Helper;
-
-    public function testFromNameReturnsExample(): void
-    {
-        $name = self::faker()->sentence;
-
-        $example = Example::fromName($name);
-
-        self::assertSame($name, $example->name());
-    }
-}
